Moved load_class to class_loader
